+++ crates/config/src/layer.rs
@@ -2,6 +2,7 @@ use crate::config::Config;
 use crate::source::Source;
 use serde::Serialize;
 
+/// A layer of configuration that was loaded, and used to create the final state.
 #[derive(Clone, Debug, Eq, PartialEq, Serialize)]
 pub struct Layer<T: Config> {
     /// The partial configuration that was loaded.